requirements.txt
opencv-python~=4.9.0.80
numpy~=1.26.4
gitdb~=4.0.11
pillow~=10.2.0
pip~=24.0
wheel~=0.43.0
Jinja2~=3.1.3
filelock~=3.13.1
pytz~=2024.1
torch~=2.2.1
thop~=0.1.1.post2209072238
requests~=2.31.0
PyYAML~=6.0.1
smmap~=5.0.1
setuptools~=69.2.0
scipy~=1.12.0
packaging~=24.0
matplotlib~=3.8.3
mpmath~=1.3.0
sympy~=1.12
tqdm~=4.66.2
torchvision~=0.17.1
psutil~=5.9.8
urllib3~=2.2.1
fsspec~=2024.3.0
MarkupSafe~=2.1.5
pandas~=2.2.1
python-dateutil~=2.9.0.post0
certifi~=2024.2.2
py-cpuinfo~=9.0.0
seaborn~=0.13.2
cycler~=0.12.1
six~=1.16.0
networkx~=3.2.1
idna~=3.6
contourpy~=1.2.0
fonttools~=4.50.0
pyparsing~=3.1.2
kiwisolver~=1.4.5
ultralytics~=8.1.29